server/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
pydantic==2.6.1
python-dotenv==1.0.1

openai==1.12.0
google-genai==1.12.1
httpx==0.28.1
numpy==1.26.4
rank-bm25==0.2.2
qdrant-client==1.15.1
sentence-transformers==2.7.0
einops==0.8.1
tree-sitter==0.21.3
tree-sitter-languages==1.10.2

ragas==0.1.10
datasets==2.18.0
pandas==2.2.0
